[01199] Move resolveColor, CHROMATIC_COLORS, and DEFAULT_SEMANTIC_COLORS to theme.ts
