+++ src/frontend/src/lib/theme.ts
@@ -58,6 +58,52 @@ export interface ThemeColors {
   radius: string;
 }
 
+export const CHROMATIC_COLORS: Record<string, string> = {
+  black: "#000000",
+  white: "#ffffff",
+  slate: "#64748b",
+  gray: "#6b7280",
+  zinc: "#71717a",
+  neutral: "#737373",
+  stone: "#78716c",
+  red: "#ef4444",
+  orange: "#f97316",
+  amber: "#f59e0b",
+  yellow: "#eab308",
+  lime: "#84cc16",
+  green: "#22c55e",
+  emerald: "#10b981",
+  teal: "#14b8a6",
+  cyan: "#06b6d4",
+  sky: "#0ea5e9",
+  blue: "#3b82f6",
+  indigo: "#6366f1",
+  violet: "#8b5cf6",
+  purple: "#a855f7",
+  fuchsia: "#d946ef",
+  pink: "#ec4899",
+  rose: "#f43f5e",
+};
+
+export const DEFAULT_SEMANTIC_COLORS: Record<string, string> = {
+  primary: "#3b82f6",
+  secondary: "#6b7280",
+  destructive: "#ef4444",
+  muted: "#6b7280",
+  foreground: "#000000",
+  background: "#ffffff",
+};
+
+export function resolveColor(
+  color: string | undefined,
+  fallback: string,
+  colorMap: Record<string, string>,
+): string {
+  if (!color) return fallback;
+  if (color.startsWith("#") || color.startsWith("rgb")) return color;
+  return colorMap[color.toLowerCase()] ?? fallback;
+}
+
 export function getThemeColors(): ThemeColors {
   return {
     background: getCSSVariable("--background"),
